requirements.txt
PyJWT~=2.8.0
cryptography~=41.0.2
fastapi~=0.99.1
aiosqlite~=0.17.0
pydantic~=1.10.11